requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
cattrs==25.3.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
espn-api==0.45.1
fastapi==0.119.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.4
pandas==2.3.3
platformdirs==4.5.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
requests-cache==1.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.2.3
uvicorn==0.37.0
